Update STACKS-2D.py
For now, the topmost stack is oscillating
No. of initial stacks has been increased to 9~10 I guess to fill up atleast half of the screen

Next step would be setting up the stack array in order to display the older stacks. Kinda updating the stack array thing because for now it is all static. We need to set it up so that it leaves a trail of all the older stacks with their positions